Improved  experiment module

+++ experiment.py
@@ -1,5 +1,6 @@
 from node2vec import *
 from sklearn import neighbors
+from credentials import *
 import sys    # sys.setdefaultencoding is cancelled by site.py
 reload(sys)    # to re-enable sys.setdefaultencoding()
 sys.setdefaultencoding('utf-8')
@@ -15,7 +16,7 @@ class experiment:
         self.label = label
         self.trainset_p = trainset_p
         self.param = param
-        self.p = figure(plot_width=1000, plot_height=600)    
+        self.p = figure(plot_width=600, plot_height=400)    
         self.ratiosf = {}
         self.r_desv = {}
         self.n_desv = {}
@@ -24,8 +25,63 @@ class experiment:
         pal = pallete("db")
         X = []
         Y = []
+        i = 1
+        if self.param == "ns":
+            k = 3
+        if self.param == "l":
+            k = 3
+        if self.param == "ndim":
+            k = 3
+        if self.param == "k":
+            k = i
+        val = i + (jump * i)
+        if i == 1:
+            val = i
+        else:
+            val = i * jump            
+        if not os.path.exists("models/ntype_prediction" + self.bd +"ts"+str(self.trainset_p)+self.param+str(val)+"k"+str(k)+".p"):
+            if self.param == "ns":
+                n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,val,200,6,self.mode,[])
+                k = 3
+            if self.param == "l":
+                n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,1000000,200,val,self.mode,[])
+                k = 3
+            if self.param == "ndim":
+                n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,1000000,val,6,self.mode,[])
+                k = 3
+            if self.param == "k":
+                n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,1000000,200,6,self.mode,[])
+                k = i
+            n2v.connectZODB()
+            n2v.learn(self.mode,self.trainset_p,False)
+            #k-neighbors for each node
+            total = 0
+            right = 0
+            for t in n2v.n_types:
+                for n in  n2v.n_types[t]:
+                    if n in n2v.w2v and random.random() < self.trainset_p:
+                        votes = []
+                        d = 10
+                        while len(votes) < k:
+                            votes = []
+                            sim = n2v.w2v.most_similar(positive = [n],topn=d)
+                            for idx,s in enumerate(sim):
+                                if n2v.ntype(s[0]) != None: 
+                                    votes.append(n2v.ntype(s[0]))
+                            d += 10          
+                        if t == max(set(votes), key=votes.count):
+                            right += 1
+                        total += 1
 
-
+            result = float(right)/float(total)
+            f = open( "models/ntype_prediction" + self.bd +"ts"+str(self.trainset_p)+self.param+str(val)+"k"+str(k)+".p", "w" )
+            n2v.disconnectZODB()
+            pickle.dump(result,f)
+        else:
+            f = open( "models/ntype_prediction" + self.bd +"ts"+str(self.trainset_p)+self.param+str(val)+"k"+str(k)+".p", "r" )
+            result = pickle.load(f)
+        X.append(val)
+        Y.append(result)
         for i in range(a,b+1):
             if self.param == "ns":
                 k = 3
@@ -35,20 +91,22 @@ class experiment:
                 k = 3
             if self.param == "k":
                 k = i
-            if not os.path.exists("models/ntype_prediction" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)+".p"):
+            val = i * jump            
+            if not os.path.exists("models/ntype_prediction" + self.bd +"ts"+str(self.trainset_p)+self.param+str(val)+"k"+str(k)+".p"):
                 if self.param == "ns":
-                    n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,i*jump,200,6,self.mode,[])
+                    n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,val,200,6,self.mode,[])
                     k = 3
                 if self.param == "l":
-                    n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,1000000,200,i*jump,self.mode,[])
+                    n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,1000000,200,val,self.mode,[])
                     k = 3
                 if self.param == "ndim":
-                    n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,1000000,i*jump,6,self.mode,[])
+                    n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,1000000,val,6,self.mode,[])
                     k = 3
                 if self.param == "k":
                     n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,1000000,200,6,self.mode,[])
                     k = i
-                n2v.learn(self.mode,self.trainset_p)
+                n2v.connectZODB()
+                n2v.learn(self.mode,self.trainset_p,False)
                 #k-neighbors for each node
                 total = 0
                 right = 0
@@ -61,8 +119,7 @@ class experiment:
                                 votes = []
                                 sim = n2v.w2v.most_similar(positive = [n],topn=d)
                                 for idx,s in enumerate(sim):
-                                    if n2v.ntype(s[0]) < > None: 
-i                                                                                                          :
+                                    if n2v.ntype(s[0]) != None: 
                                         votes.append(n2v.ntype(s[0]))
                                 d += 10          
                             if t == max(set(votes), key=votes.count):
@@ -70,12 +127,13 @@ i                                                                               
                             total += 1
 
                 result = float(right)/float(total)
-                f = open( "models/ntype_prediction" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)+".p", "w" )
+                f = open( "models/ntype_prediction" + self.bd +"ts"+str(self.trainset_p)+self.param+str(val)+"k"+str(k)+".p", "w" )
+                n2v.disconnectZODB()
                 pickle.dump(result,f)
             else:
-                f = open( "models/ntype_prediction" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)+".p", "r" )
+                f = open( "models/ntype_prediction" + self.bd +"ts"+str(self.trainset_p)+self.param+str(val)+"k"+str(k)+".p", "r" )
                 result = pickle.load(f)
-            X.append(i*jump)
+            X.append(val)
             Y.append(result)
         self.p.line(X, Y, color=pal[1],legend="ICH",line_width=1.5)
         self.p.legend.background_fill_alpha = 0.5
@@ -96,7 +154,9 @@ i                                                                               
                 k = 3
             if self.param == "k":
                 k = i
+
             if not os.path.exists("models/ltype_prediction" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)+".p"):
+                print "models/ltype_prediction" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)+".p"
                 if self.param == "ns":
                     n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,i*jump,200,6,self.mode,[])
                     k = 3
@@ -110,14 +170,14 @@ i                                                                               
                     n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,1000000,200,6,self.mode,[])
                     k = i
                 n2v.connectZODB()
-                n2v.learn(self.mode,self.trainset_p)
+                n2v.learn(self.mode,self.trainset_p,False)
                 #k-neighbors for each link
                 total = 0
                 right = 0
                 X,Y,T =[],[],[]
                 for t in n2v.r_types:
                     for r in n2v.r_types[t]:
-                        if random.random() < self.trainset_p:
+                        if random.random() < self.trainset_p+0.1:
                             T.append([r["v"],t])
                             
                         else:   
@@ -147,33 +207,77 @@ i                                                                               
 
     def link_prediction(self,traversals,a,b,jump):
         pal = pallete("links")
+        self.ratiosf = {}
+        self.n_desv = {}
+        self.r_desv = {}
         for i in range(a,b):
+
+            if self.param == "ns":
+                k = 3
+            if self.param == "l":
+                k = 3
+            if self.param == "ndim":
+                k = 3
+            if self.param == "k":
+                k = i
             if self.param == "ns":
                 n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,i*jump,200,6,self.mode,traversals)
             if self.param == "l":
                 n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,250000,200,i*jump,self.mode,traversals)
             if self.param == "ndim":
                 n2v = node2vec(self.bd,self.port,self.user,self.pss,self.label,250000,i*jump,6,self.mode,traversals)
-            n2v.delete_rels(self.trainset_p)
-            n2v.learn(self.mode,self.trainset_p)
-            for r in n2v.r_deleted:
-                print r
-                if r[0] == '-':
-                    con = neo4j.CypherQuery(n2v.graph_db, "MATCH (a)"+r+"(b) WHERE labels(a) <> [] AND labels(b) <> [] RETURN DISTINCT head(labels(a)) AS This, type(r) as To, head(labels(b)) AS That LIMIT 5").execute()
-                else:
-                    con = neo4j.CypherQuery(n2v.graph_db, "MATCH (a)-[r:"+r+"]->(b) WHERE labels(a) <> [] AND labels(b) <> [] RETURN DISTINCT head(labels(a)) AS This, type(r) as To, head(labels(b)) AS That LIMIT 5").execute()
-                if not r in self.ratiosf:
-                    self.ratiosf[r] = []
-                self.ratiosf[r].append(n2v.aciertos_rel(r,self.label,True,str(i*jump)+self.param+str(self.trainset_p)))
-                if not r in self.r_desv:
-                    self.r_desv[r] = []
-                self.r_desv[r].append(n2v.r_desv[r])
-                if not r in self.n_desv:
-                    self.n_desv[r] = []
-                print "MATCH (a)-[r:"+r+"]->(b) WHERE labels(a) <> [] AND labels(b) <> [] RETURN DISTINCT head(labels(a)) AS This, type(r) as To, head(labels(b)) AS That LIMIT 5"
-                self.n_desv[r].append(n2v.n_types_d[con[0]["That"]])
-                ndesv = []
+            n2v.connectZODB()
+            if "lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k) not in n2v.root:
+                print "entrando"
+                n2v.learn(self.mode,self.trainset_p,False)
+                n2v.delete_rels(self.trainset_p)
+                n2v.learn(self.mode,self.trainset_p,True)
+                n2v.n_analysis()
+                n2v.r_analysis()
+
+                n2v.root["lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)] = PersistentDict()
+                n2v.root["lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)]["ratiosf"] = PersistentDict()
+                n2v.root["lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)]["r_desv"] = PersistentDict()
+                n2v.root["lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)]["n_desv"] = PersistentDict()
+                total = 0
+                suma = 0
+                for r in n2v.r_deleted:
+                    print r
+                    if r[0] == '-':
+                        con = neo4j.CypherQuery(n2v.graph_db, "MATCH (a)"+r+"(b) WHERE labels(a) <> [] AND labels(b) <> [] RETURN DISTINCT head(labels(a)) AS This, type(r) as To, head(labels(b)) AS That LIMIT 5").execute()
+                    else:
+                        con = neo4j.CypherQuery(n2v.graph_db, "MATCH (a)-[r:"+r+"]->(b) WHERE labels(a) <> [] AND labels(b) <> [] RETURN DISTINCT head(labels(a)) AS This, type(r) as To, head(labels(b)) AS That LIMIT 5").execute()
+                    if not r in self.ratiosf:
+                        self.ratiosf[r] = []
+                    aciertos = n2v.aciertos_rel(r,self.label,True,str(i*jump)+self.param+str(self.trainset_p))
+                    self.ratiosf[r].append(aciertos)
+                    n2v.root["lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)]["ratiosf"][r] = aciertos
+                    if not r in self.r_desv:
+                        self.r_desv[r] = []
+                    self.r_desv[r].append(n2v.r_desv[r])
+                    n2v.root["lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)]["r_desv"][r] = n2v.r_desv[r]
+                    if not r in self.n_desv:
+                        self.n_desv[r] = []
+                    print "MATCH (a)-[r:"+r+"]->(b) WHERE labels(a) <> [] AND labels(b) <> [] RETURN DISTINCT head(labels(a)) AS This, type(r) as To, head(labels(b)) AS That LIMIT 5"
+                    self.n_desv[r].append(n2v.n_types_d[con[0]["That"]])
+                    n2v.root["lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)]["n_desv"][r]  = n2v.n_types_d[con[0]["That"]]
+                    ndesv = []
+            else:
+                for r in n2v.root["lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)]["ratiosf"]:
+                    if not r in self.ratiosf:
+                        self.ratiosf[r] = []
+                    self.ratiosf[r].append(n2v.root["lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)]["ratiosf"][r])    
+                    if not r in self.r_desv:
+                        self.r_desv[r] = []
+                    self.r_desv[r].append(n2v.root["lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)]["r_desv"][r])  
+                    if not r in self.n_desv:
+                        self.n_desv[r] = []  
+                    self.n_desv[r].append(n2v.root["lp" + self.bd +"ts"+str(self.trainset_p)+self.param+str(i*jump)+"k"+str(k)]["n_desv"][r])    
+            n2v.disconnectZODB()
         print self.ratiosf
+        print self.r_desv
+        print self.n_desv
+        
         for j,r in enumerate(self.ratiosf):
             if r <> "HERRAMIENTA" and r <>  "PARROQUIA_LOC" and r <> "CANTON_LOC" and r <> "DSUBAMBITO_PERTENECE": 
                 x = []
@@ -184,8 +288,8 @@ i                                                                               
                     print "ja"
                     x.append(i*jump)
                     ratios.append(ratio)
-                    rdesv.append(self.r_desv[r][i]*10)
-                    ndesv.append(self.n_desv[r][i]*10)
+                    rdesv.append(self.r_desv[r][i])
+                    ndesv.append(self.n_desv[r][i])
                 print ratios
                 print x
                 self.p.line(x, ratios, color=pal[j],legend=r+"(LPR)",line_width=1.5)
